% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank Dataset}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{...1}{Row number}
\item{arrival_time}{Customer arrival time}
\item{service_time}{Service duration}
}
}
\source{
Generated for Assignment 2
}
\usage{
bank
}
\description{
Simulated arrival and service times for bank customers.
}
\keyword{datasets}
