% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Queue simulation}
\usage{
Multiserver(arrival_time, service_time, servers = 3)
}
\arguments{
\item{arrival_time}{Numeric vector of arrival times}

\item{service_time}{Numeric vector of service times}

\item{servers}{Number of servers}
}
\value{
tibble result
}
\description{
Queue simulation
}
